+++ app/Http/Controllers/PaymentMethodController.php
@@ -146,6 +146,9 @@ class PaymentMethodController extends Controller
     {
        DB::beginTransaction();
         try {
+            if(!auth()->user()->isAdmin()){
+                return Helper::inValidRequest('You are not unauthorized to perform this operation.', 'Unauthorized Access!', 400);
+            }
             $data = $paymentmethod->delete();
             DB::commit();
             return Helper::validRequest(["success" => $data], 'Item deleted successfully', 200);
